requirements.txt
opencv-python
pygetwindow
pyautogui
Appium-Python-Client
pytest